+++ src/element/shape/Circle.cpp
@@ -1,7 +1,7 @@
 #include "Circle.h"
 
-Circle::Circle(Vector2D<double> position, const SVGColor& fillColor, const SVGColor& strokeColor, double strokeWidth, 
-	           double radius) : Ellipse(position, fillColor, strokeColor, strokeWidth, Vector2D<double>(radius, radius)) {};
+Circle::Circle(Vector2D<float> position, const SVGColor& fillColor, const SVGColor& strokeColor, float strokeWidth, 
+	           float radius) : Ellipse(position, fillColor, strokeColor, strokeWidth, Vector2D<float>(radius, radius)) {};
 
 ElementType Circle::getTypeName() {
 	return ElementType::Circle;
